requirements.txt
streamlit
python-dotenv
groq
PyPDF2